Kombinert iverksetting test i en større integrasjonstest.
